+++ src/components/ContentPartialsEditor/ContentPartialsEditor.test.tsx
@@ -1,15 +1,15 @@
 import { act, fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
-import { DragDropContext, DropResult } from 'react-beautiful-dnd';
+import { DragDropContext, DropResult } from '@hello-pangea/dnd';
 
 import { TEST_IDS } from '../../constants';
 import { ContentPartialsEditor } from './ContentPartialsEditor';
 
 /**
- * Mock react-beautiful-dnd
+ * Mock @hello-pangea/dnd
  */
-jest.mock('react-beautiful-dnd', () => ({
-  ...jest.requireActual('react-beautiful-dnd'),
+jest.mock('@hello-pangea/dnd', () => ({
+  ...jest.requireActual('@hello-pangea/dnd'),
   DragDropContext: jest.fn(({ children }) => children),
   Droppable: jest.fn(({ children }) => children({})),
   Draggable: jest.fn(({ children }) =>